requirements.txt
streamlit
pandas
pycaret.clustering
plotly.express
numpy
python-dotenv
openai